feat: setup build script
