+++ .gitignore
@@ -4,3 +4,5 @@ uv.lock
 .python-version
 build
 *.egg-info
+
+.cache
