+++ include/export.h
@@ -1,0 +1,5 @@
+namespace maya_usd_export{
+    class Export {
+    public:
+    };
+}
